pyproject.toml
[tool.poetry]
name = "miniscope-wired-interconnect-characterization"
description = "Characterization data, processing, and analysis for Miniscope wired interconnects"
version = "0.1.0"
authors = ["Daniel Aharoni"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.3"
pytest-cov = "^4.1"

[tool.poetry.scripts]
miniscope-char = "src.cli.main:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "I", "UP"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
